added entrypoint script which compiles a flutter app
The container will compile a flutter app, which is mounted to C:\src .
After building the app, the result is copied to C:\src\build_container.
Two options are provided by the script and can be specified on the commandline:

-example
Build the example app of a flutter package mounted at C:\src.
- msix
Package the built app as msix. The app must have the msix package listed
as dependency.